requirements.txt
streamlit
groq
python-dotenv
google-api-python-client
requests